requirements.txt
easyocr
gspread
oauth2client
urllib3==1.26.15
google-genai
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib
mcp
mcp[cli]